pyproject.toml
[tool.poetry]
name = "openhands-package"
version = "0.1.0"
description = "Testing The OpenHands Package"
authors = ["Matt Barlow <matt.barlow@stratusgrid.com>"]

[tool.poetry.dependencies]
python = "^3.11"
asyncio = "^3.4.3"
click = "^8.1.7"
inquirerpy = "^0.3.4"
openhands-ai = "^0.8.3"
python-docx = "^1.1.2"
pypdf2 = "^3.0.1"
python-pptx = "^1.0.2"
pylatexenc = "^2.10"
tornado = "^6.4.1"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
openhands-package = "openhands_package.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
